% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{apexchart_dep}
\alias{apexchart_dep}
\title{Remove the old dependency functions since we're using local files now}
\usage{
apexchart_dep()
}
\description{
Remove the old dependency functions since we're using local files now
}
