% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_plots.R
\name{district_summary_table}
\alias{district_summary_table}
\title{Create a styled district summary table}
\usage{
district_summary_table(data, color_pal = c("#ffffd9", "#c7e9b4", "#41b6c4"))
}
\arguments{
\item{data}{A dataframe containing the summary data}

\item{color_pal}{Color palette for the background gradient (default: c("#f7fbff", "#08306b"))}
}
\value{
A reactable object
}
\description{
Create a styled district summary table
}
