% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{data}
\name{.app_env}
\alias{.app_env}
\title{Application parameters environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.app_env
}
\description{
Application parameters environment
}
\keyword{internal}
