% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hex_map.R
\name{hexagon_map}
\alias{hexagon_map}
\title{Create a hexagon heatmap using deckgl}
\usage{
hexagon_map(
  data = NULL,
  tracks_data = peskas.malawi.portal::tracks_df,
  radius = NULL,
  elevation_scale = 80,
  pitch = 45,
  zoom = 8.5,
  color_by = NULL,
  color_from = "#edf8b1",
  color_to = "#41b6c4",
  default_color = "#41b6c4",
  path_opacity = 0.5,
  path_width = 0.75
)
}
\arguments{
\item{data}{A dataframe containing the map data}

\item{tracks_data}{Optional dataframe containing track data}

\item{radius}{Numeric. Radius of hexagons in meters}

\item{elevation_scale}{Numeric. Elevation scale for 3D effect}

\item{pitch}{Numeric. Map pitch in degrees}

\item{zoom}{Numeric. Initial zoom level}

\item{color_by}{Character. Column name to use for coloring}

\item{color_from}{Character. Starting color in hex format}

\item{color_to}{Character. Ending color in hex format}

\item{default_color}{Character. Default color in hex format}

\item{path_opacity}{Numeric. Opacity of paths}

\item{path_width}{Numeric. Width of paths}
}
\value{
A deckgl object
}
\description{
Generates an interactive hexagon heatmap visualization.
}
