% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peskas.malawi.portal-package.R
\docType{package}
\name{peskas.malawi.portal-package}
\alias{peskas.malawi.portal}
\alias{peskas.malawi.portal-package}
\title{peskas.malawi.portal: Peskas malawi ssf dashboard}
\description{
A web portal displaying data and insights from small fisheries in Malawi.
}
\author{
\strong{Maintainer}: Lorenzo Longobardi \email{l.longobardi@cgiar.org} (\href{https://orcid.org/0000-0003-3126-7341}{ORCID})

Other contributors:
\itemize{
  \item WorldFish [copyright holder]
}

}
\keyword{internal}
