% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_validation.R
\name{tab_validation_content}
\alias{tab_validation_content}
\title{Generate content for the validation tab}
\usage{
tab_validation_content()
}
\value{
A tagList containing the structured content for the validation tab
}
\description{
This function creates the layout and content for the validation tab of the dashboard.
}
