% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_validation.R
\name{validation_modal}
\alias{validation_modal}
\title{Create validation modal}
\usage{
validation_modal()
}
\value{
A shiny.tag object containing the validation modal
}
\description{
Create validation modal
}
