% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R.R
\docType{data}
\name{validation_table}
\alias{validation_table}
\title{Validation table data}
\format{
A data frame with submission information
}
\usage{
validation_table
}
\description{
Data for validation interface
}
\keyword{datasets}
