% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_plots.R
\name{hexagon_map}
\alias{hexagon_map}
\title{Generate a hexagon heatmap using deckgl}
\usage{
hexagon_map(data, radius = 3000, elevation_scale = 80, pitch = 45, zoom = 6.5)
}
\arguments{
\item{data}{A dataframe containing the map data with 'lat' and 'lon' columns}

\item{radius}{Radius of hexagons in meters (default: 3000)}

\item{elevation_scale}{Elevation scale for 3D effect (default: 80)}

\item{pitch}{Map pitch in degrees (default: 45)}

\item{zoom}{Initial zoom level (default: 7)}
}
\description{
Generate a hexagon heatmap using deckgl
}
